backend/pyproject.toml
[project]
name = "auto-inquiry-backend"
version = "0.1.0"
description = "Backend API for Auto Inquiry Form Submitter"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.20",
    "celery[redis]==5.4.0",
    "redis==5.2.1",
    "playwright==1.50.0",
    "boto3==1.36.8",
    "python-dotenv==1.0.1",
    "email-validator==2.2.0",
    "pluggy>=1.6.0",
    "pytest-dependency>=0.6.0",
    "croniter>=6.0.0",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.1",
    "ruff==0.9.1",
    "pyright==1.1.391",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
